Fix currency display: Convert all prices from paise to rupees across all modules
- Updated all price displays to show rupees directly (removed /100 divisions)
- Changed field names: costPaise/quotedCostPaise to cost/quotedCost
- Updated validators to accept decimal prices instead of integers
- Fixed price calculations in products, orders, purchase orders, reports, analytics
- Updated customer-facing pages, supplier pages, and public store pages
- Added currency migration scripts and tax fields migration
- Updated all API endpoints to handle Decimal types correctly
- Added formatCurrency and decimalToNumber helpers throughout the codebase

+++ apps/web/src/lib/validators.ts
@@ -5,8 +5,8 @@ export const createProductSchema = z.object({
   title: z.string().min(1, 'Title is required'),
   description: z.string().optional(),
   sku: z.string().optional(),
-  price: z.number().int().min(0, 'Price must be non-negative'),
-  costPrice: z.number().int().min(0).optional(),
+  price: z.number().min(0, 'Price must be non-negative'), // Now in rupees (decimal)
+  costPrice: z.number().min(0).optional(), // Now in rupees (decimal)
   stock: z.number().int().min(0, 'Stock must be non-negative'),
   reorderPoint: z.number().int().min(0, 'Reorder point must be non-negative'),
   reorderQty: z.number().int().min(0, 'Reorder quantity must be non-negative'),
@@ -89,8 +89,8 @@ export const csvProductSchema = z.object({
   title: z.string().min(1, 'Title is required'),
   description: z.string().optional(),
   sku: z.string().optional(),
-  price: z.number().int().min(0, 'Price must be non-negative'),
-  costPrice: z.number().int().min(0).optional(),
+  price: z.number().min(0, 'Price must be non-negative'), // Now in rupees (decimal)
+  costPrice: z.number().min(0).optional(), // Now in rupees (decimal)
   stock: z.number().int().min(0, 'Stock must be non-negative'),
   reorderPoint: z.number().int().min(0, 'Reorder point must be non-negative'),
   reorderQty: z.number().int().min(0, 'Reorder quantity must be non-negative'),
